Move the chat API out of beta. Save chat sessions at 15min intervals
