+++ src/khoj/routers/api.py
@@ -10,6 +10,7 @@ from fastapi import HTTPException
 
 # Internal Packages
 from khoj.configure import configure_processor, configure_search
+from khoj.processor.conversation.gpt import converse, message_to_log, message_to_prompt
 from khoj.search_type import image_search, text_search
 from khoj.utils.helpers import timer
 from khoj.utils.rawconfig import FullConfig, SearchResponse
@@ -183,3 +184,40 @@ def update(t: Optional[SearchType] = None, force: Optional[bool] = False):
         logger.info("📬 Processor reconfigured via API")
 
     return {"status": "ok", "message": "khoj reloaded"}
+
+
+@api.get("/chat")
+def chat(q: Optional[str] = None):
+    # Initialize Variables
+    api_key = state.processor_config.conversation.openai_api_key
+
+    # Load Conversation History
+    chat_session = state.processor_config.conversation.chat_session
+    meta_log = state.processor_config.conversation.meta_log
+
+    # If user query is empty, return chat history
+    if not q:
+        if meta_log.get("chat"):
+            return {"status": "ok", "response": meta_log["chat"]}
+        else:
+            return {"status": "ok", "response": []}
+
+    # Collate context for GPT
+    result_list = search(q, n=2, r=True, score_threshold=0, dedupe=False)
+    collated_result = "\n\n".join([f"# {item.additional['compiled']}" for item in result_list])
+    logger.debug(f"Reference Context:\n{collated_result}")
+
+    try:
+        gpt_response = converse(collated_result, q, meta_log, api_key=api_key)
+        status = "ok"
+    except Exception as e:
+        gpt_response = str(e)
+        status = "error"
+
+    # Update Conversation History
+    state.processor_config.conversation.chat_session = message_to_prompt(q, chat_session, gpt_message=gpt_response)
+    state.processor_config.conversation.meta_log["chat"] = message_to_log(
+        q, gpt_response, khoj_message_metadata={"context": collated_result}, conversation_log=meta_log.get("chat", [])
+    )
+
+    return {"status": status, "response": gpt_response, "context": collated_result}
